# packages/client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bandroom.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5180/"
    }
  ],
  "paths": {
...
    "/bands/{bandId}/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "Comments on an event or demo version",
        "parameters": [
          {
            "name": "targetType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Comment (optionally at a timestamp)",
        "parameters": [
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bands/{bandId}/comments/{commentId}": {
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "Delete own comment (or as admin)",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentResponse": {
        "required": [
          "id",
          "targetType",
          "targetId",
          "authorMembershipId",
          "authorName",
          "body",
          "atSeconds",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "targetType": {
            "$ref": "#/components/schemas/CommentTarget"
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "authorMembershipId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "atSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      },
      "CommentTarget": {
        "enum": [
          "event",
          "demoVersion"
        ]
      },
...
      "CreateCommentRequest": {
        "required": [
          "targetType",
          "targetId",
          "body"
        ],
        "type": "object",
        "properties": {
          "targetType": {
            "$ref": "#/components/schemas/CommentTarget"
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "atSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Instant": {
        "type": "string",
        "format": "date-time"
      },
...
    }
  },
...
}